refactor: split ui/app.rs into modules, reduce monolithical chunk
