+++ src/ui/app.rs
@@ -138,6 +138,7 @@ mod presets;
 mod priority_control_pages;
 mod process_list_page;
 mod process_policies;
+mod processor_power;
 mod settings_pages;
 mod widgets;
 
@@ -284,77 +285,6 @@ const COLOR_LIGHT_BORDER: u32 = 0xc7ccd2;
 const COLOR_LIGHT_TEXT: u32 = 0x171a1d;
 const COLOR_LIGHT_MUTED: u32 = 0x565d64;
 const COLOR_LIGHT_DIM: u32 = 0x747c84;
-
-#[derive(Clone, Copy)]
-struct DropdownPlacement {
-    open_up: bool,
-    max_height: Pixels,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum ActionLogResultFilter {
-    All,
-    Applied,
-    Restored,
-    Skipped,
-    Failed,
-}
-
-impl ActionLogResultFilter {
-    const ALL: [Self; 5] = [
-        Self::All,
-        Self::Applied,
-        Self::Restored,
-        Self::Skipped,
-        Self::Failed,
-    ];
-
-    fn matches(self, result: ActionLogResult) -> bool {
-        match self {
-            Self::All => true,
-            Self::Applied => result == ActionLogResult::Applied,
-            Self::Restored => result == ActionLogResult::Restored,
-            Self::Skipped => result == ActionLogResult::Skipped,
-            Self::Failed => result == ActionLogResult::Failed,
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum ActionLogFeatureFilter {
-    All,
-    Feature(ActionLogFeature),
-}
-
-impl ActionLogFeatureFilter {
-    const ALL: [Self; 16] = [
-        Self::All,
-        Self::Feature(ActionLogFeature::AppSuspension),
-        Self::Feature(ActionLogFeature::BackgroundCpuRestriction),
-        Self::Feature(ActionLogFeature::CoreSteering),
-        Self::Feature(ActionLogFeature::BackgroundEfficiency),
-        Self::Feature(ActionLogFeature::CoreLimiter),
-        Self::Feature(ActionLogFeature::ByRunningApp),
-        Self::Feature(ActionLogFeature::WorkloadEngine),
-        Self::Feature(ActionLogFeature::ProcessPriority),
-        Self::Feature(ActionLogFeature::ThreadPriority),
-        Self::Feature(ActionLogFeature::DynamicPriorityBoost),
-        Self::Feature(ActionLogFeature::IoPriority),
-        Self::Feature(ActionLogFeature::GpuPriority),
-        Self::Feature(ActionLogFeature::MemoryPriority),
-        Self::Feature(ActionLogFeature::MemoryTrim),
-        Self::Feature(ActionLogFeature::TimerResolution),
-    ];
-
-    fn matches(self, feature: ActionLogFeature) -> bool {
-        match self {
-            Self::All => true,
-            Self::Feature(filter_feature) => filter_feature == feature,
-        }
-    }
-}
-
-const ACTION_LOG_PAGE_SIZE: usize = 15;
 
 const ACCENT_PALETTE: [u32; 48] = [
     0xa7e957, 0xc7f36d, 0x8fd14f, 0x65b741, 0x3f8f34, 0x2f6f34, 0xd8c75b, 0xffc857, 0xe0a93a,
@@ -792,252 +722,6 @@ enum Win32PrioritySeparationField {
 #[derive(Clone, Copy)]
 struct Win32PrioritySeparationFieldOption {
     bits: u32,
-}
-
-impl UiInputs {
-    fn new(
-        window: &mut Window,
-        cx: &mut Context<WinderustApp>,
-        settings: &Settings,
-        processor_power_values: ProcessorPowerAcDcValues,
-    ) -> Self {
-        let processor_power_values = processor_power_values.normalized();
-        Self {
-            dashboard_search: make_input(window, cx, "", &t!("home.search_placeholder")),
-            by_cpu_load_rule_names: settings
-                .by_cpu_load
-                .rules
-                .iter()
-                .map(|rule| make_input(window, cx, &rule.name, &t!("common.rule_name")))
-                .collect(),
-            cpu_rule_thresholds: settings
-                .by_cpu_load
-                .rules
-                .iter()
-                .map(|rule| make_percent_slider(cx, rule.threshold_percent as u64))
-                .collect(),
-            cpu_rule_upper_thresholds: settings
-                .by_cpu_load
-                .rules
-                .iter()
-                .map(|rule| {
-                    make_percent_slider(cx, rule.upper_threshold_percent.unwrap_or(100) as u64)
-                })
-                .collect(),
-            by_time_rule_names: settings
-                .by_time
-                .rules
-                .iter()
-                .map(|rule| make_input(window, cx, &rule.name, &t!("common.rule_name")))
-                .collect(),
-            schedule_start_times: settings
-                .by_time
-                .rules
-                .iter()
-                .map(|rule| make_input(window, cx, &rule.start_time, "HH:MM"))
-                .collect(),
-            schedule_end_times: settings
-                .by_time
-                .rules
-                .iter()
-                .map(|rule| make_input(window, cx, &rule.end_time, "HH:MM"))
-                .collect(),
-            foreground_rule_names: settings
-                .by_foreground
-                .rules
-                .iter()
-                .map(|rule| make_input(window, cx, &rule.name, &t!("common.rule_name")))
-                .collect(),
-            foreground_rule_processes: settings
-                .by_foreground
-                .rules
-                .iter()
-                .map(|rule| make_input(window, cx, &rule.process_name, "process.exe"))
-                .collect(),
-            foreground_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            background_efficiency_process: make_input(
-                window,
-                cx,
-                "",
-                &t!("common.search_running_apps"),
-            ),
-            background_cpu_exclusion: make_input(window, cx, "", &t!("common.search_running_apps")),
-            memory_trim_exclusion: make_input(window, cx, "", &t!("common.search_running_apps")),
-            app_suspension_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            core_limiter_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            performance_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            core_steering_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            workload_engine_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            process_priority_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            thread_priority_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            dynamic_priority_boost_process: make_input(
-                window,
-                cx,
-                "",
-                &t!("common.search_running_apps"),
-            ),
-            io_priority_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            gpu_priority_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            memory_priority_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            timer_resolution_process: make_input(window, cx, "", &t!("common.search_running_apps")),
-            numeric_value: make_input(window, cx, "", "Value"),
-            activity_idle_timeout: make_range_slider(
-                cx,
-                settings.by_activity.idle_timeout_seconds,
-                ACTIVITY_IDLE_TIMEOUT_MIN_SECONDS,
-                ACTIVITY_IDLE_TIMEOUT_MAX_SECONDS,
-                1,
-            ),
-            activity_check_interval: make_range_slider(
-                cx,
-                settings.general.check_interval_ms,
-                ACTIVITY_CHECK_INTERVAL_MIN_MS,
-                ACTIVITY_CHECK_INTERVAL_MAX_MS,
-                ACTIVITY_CHECK_INTERVAL_STEP_MS,
-            ),
-            processor_power_ac_core_parking_min: make_processor_power_slider(
-                cx,
-                processor_power_values.ac.core_parking_min as u64,
-            ),
-            processor_power_ac_performance_min: make_processor_power_slider(
-                cx,
-                processor_power_values.ac.performance_min as u64,
-            ),
-            processor_power_ac_performance_max: make_processor_power_slider(
-                cx,
-                processor_power_values.ac.performance_max as u64,
-            ),
-            processor_power_ac_boost_policy: make_processor_power_slider(
-                cx,
-                processor_power_values.ac.boost_policy as u64,
-            ),
-            processor_power_dc_core_parking_min: make_processor_power_slider(
-                cx,
-                processor_power_values.dc.core_parking_min as u64,
-            ),
-            processor_power_dc_performance_min: make_processor_power_slider(
-                cx,
-                processor_power_values.dc.performance_min as u64,
-            ),
-            processor_power_dc_performance_max: make_processor_power_slider(
-                cx,
-                processor_power_values.dc.performance_max as u64,
-            ),
-            processor_power_dc_boost_policy: make_processor_power_slider(
-                cx,
-                processor_power_values.dc.boost_policy as u64,
-            ),
-        }
-    }
-
-    fn ensure_for_settings(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<WinderustApp>,
-        settings: &Settings,
-    ) {
-        sync_input_vec(
-            &mut self.by_cpu_load_rule_names,
-            settings.by_cpu_load.rules.len(),
-            window,
-            cx,
-            |index| settings.by_cpu_load.rules[index].name.clone(),
-            &t!("common.rule_name"),
-        );
-        sync_slider_vec(
-            &mut self.cpu_rule_thresholds,
-            settings.by_cpu_load.rules.len(),
-            cx,
-            |index| settings.by_cpu_load.rules[index].threshold_percent as u64,
-        );
-        sync_slider_vec(
-            &mut self.cpu_rule_upper_thresholds,
-            settings.by_cpu_load.rules.len(),
-            cx,
-            |index| {
-                settings.by_cpu_load.rules[index]
-                    .upper_threshold_percent
-                    .unwrap_or(100) as u64
-            },
-        );
-        sync_input_vec(
-            &mut self.by_time_rule_names,
-            settings.by_time.rules.len(),
-            window,
-            cx,
-            |index| settings.by_time.rules[index].name.clone(),
-            &t!("common.rule_name"),
-        );
-        sync_input_vec(
-            &mut self.schedule_start_times,
-            settings.by_time.rules.len(),
-            window,
-            cx,
-            |index| settings.by_time.rules[index].start_time.clone(),
-            "HH:MM",
-        );
-        sync_input_vec(
-            &mut self.schedule_end_times,
-            settings.by_time.rules.len(),
-            window,
-            cx,
-            |index| settings.by_time.rules[index].end_time.clone(),
-            "HH:MM",
-        );
-        sync_input_vec(
-            &mut self.foreground_rule_names,
-            settings.by_foreground.rules.len(),
-            window,
-            cx,
-            |index| settings.by_foreground.rules[index].name.clone(),
-            &t!("common.rule_name"),
-        );
-        sync_input_vec(
-            &mut self.foreground_rule_processes,
-            settings.by_foreground.rules.len(),
-            window,
-            cx,
-            |index| settings.by_foreground.rules[index].process_name.clone(),
-            "process.exe",
-        );
-    }
-
-    fn refresh_localized_placeholders(&self, window: &mut Window, cx: &mut Context<WinderustApp>) {
-        set_input_placeholder(
-            &self.dashboard_search,
-            t!("home.search_placeholder"),
-            window,
-            cx,
-        );
-        for input in [
-            &self.foreground_process,
-            &self.background_efficiency_process,
-            &self.background_cpu_exclusion,
-            &self.memory_trim_exclusion,
-            &self.app_suspension_process,
-            &self.core_limiter_process,
-            &self.performance_process,
-            &self.core_steering_process,
-            &self.workload_engine_process,
-            &self.process_priority_process,
-            &self.thread_priority_process,
-            &self.dynamic_priority_boost_process,
-            &self.io_priority_process,
-            &self.gpu_priority_process,
-            &self.memory_priority_process,
-            &self.timer_resolution_process,
-        ] {
-            set_input_placeholder(input, t!("common.search_running_apps"), window, cx);
-        }
-        for input in self
-            .by_cpu_load_rule_names
-            .iter()
-            .chain(&self.by_time_rule_names)
-            .chain(&self.foreground_rule_names)
-        {
-            set_input_placeholder(input, t!("common.rule_name"), window, cx);
-        }
-    }
 }
 
 fn default_processor_power_values() -> ProcessorPowerAcDcValues {
@@ -1738,418 +1422,6 @@ impl WinderustApp {
             let _ = self_power::enable_adaptive_engine();
         } else {
             let _ = self_power::disable_adaptive_engine();
-        }
-    }
-
-    fn refresh_processor_power_target_plan_personality(&mut self) -> bool {
-        let personality = self
-            .processor_power_target_plan()
-            .and_then(|plan| self.power.read_plan_personality(&plan.guid).ok());
-        if self.processor_power_target_plan_personality == personality {
-            return false;
-        }
-
-        self.processor_power_target_plan_personality = personality;
-        true
-    }
-
-    fn ensure_processor_power_target_plan(&mut self) {
-        let target_still_available = self
-            .processor_power_target_plan_guid
-            .as_deref()
-            .is_some_and(|target| {
-                self.plans
-                    .iter()
-                    .any(|plan| plan.guid.eq_ignore_ascii_case(target))
-            });
-        if target_still_available {
-            return;
-        }
-
-        self.processor_power_target_plan_guid = self
-            .current_plan
-            .as_ref()
-            .or_else(|| self.plans.first())
-            .map(|plan| plan.guid.clone());
-    }
-
-    fn processor_power_target_plan(&self) -> Option<PowerPlan> {
-        self.processor_power_target_plan_guid
-            .as_deref()
-            .and_then(|target| {
-                self.plans
-                    .iter()
-                    .find(|plan| plan.guid.eq_ignore_ascii_case(target))
-            })
-            .cloned()
-            .or_else(|| self.current_plan.clone())
-    }
-
-    fn set_processor_power_target_plan(&mut self, guid: String) {
-        self.processor_power_target_plan_guid = Some(guid);
-        self.active_power_plan_picker = None;
-        self.sync_processor_power_values_from_target_plan(true);
-    }
-
-    fn set_processor_power_target_plan_option(&mut self, guid: Option<String>) {
-        if let Some(guid) = guid {
-            self.set_processor_power_target_plan(guid);
-        } else {
-            self.active_power_plan_picker = None;
-        }
-    }
-
-    fn sync_processor_power_values_from_target_plan(&mut self, force: bool) -> bool {
-        self.refresh_processor_power_target_plan_personality();
-        let Some(plan) = self.processor_power_target_plan() else {
-            self.processor_power_loaded_plan_guid = None;
-            return false;
-        };
-        let same_plan = self
-            .processor_power_loaded_plan_guid
-            .as_deref()
-            .is_some_and(|guid| guid.eq_ignore_ascii_case(&plan.guid));
-        if !force && same_plan && self.processor_power_dirty {
-            return true;
-        }
-
-        match self.power.read_processor_power_values(&plan.guid) {
-            Ok(values) => {
-                self.set_processor_power_values(values.normalized());
-                self.processor_power_loaded_plan_guid = Some(plan.guid);
-                self.processor_power_dirty = false;
-                true
-            }
-            Err(err) => {
-                self.status_message = err;
-                false
-            }
-        }
-    }
-
-    fn processor_power_values(&self) -> ProcessorPowerAcDcValues {
-        ProcessorPowerAcDcValues::new(
-            ProcessorPowerValues::new_with_boost_mode(
-                self.processor_power_ac_core_parking_min as u32,
-                self.processor_power_ac_performance_min as u32,
-                self.processor_power_ac_performance_max as u32,
-                self.processor_power_ac_boost_policy as u32,
-                self.processor_power_ac_boost_mode,
-            ),
-            ProcessorPowerValues::new_with_boost_mode(
-                self.processor_power_dc_core_parking_min as u32,
-                self.processor_power_dc_performance_min as u32,
-                self.processor_power_dc_performance_max as u32,
-                self.processor_power_dc_boost_policy as u32,
-                self.processor_power_dc_boost_mode,
-            ),
-        )
-        .normalized()
-    }
-
-    fn set_processor_power_values(&mut self, values: ProcessorPowerAcDcValues) {
-        let values = values.normalized();
-        self.processor_power_ac_core_parking_min = values.ac.core_parking_min as u64;
-        self.processor_power_ac_performance_min = values.ac.performance_min as u64;
-        self.processor_power_ac_performance_max = values.ac.performance_max as u64;
-        self.processor_power_ac_boost_policy = values.ac.boost_policy as u64;
-        self.processor_power_ac_boost_mode = values.ac.boost_mode;
-        self.processor_power_dc_core_parking_min = values.dc.core_parking_min as u64;
-        self.processor_power_dc_performance_min = values.dc.performance_min as u64;
-        self.processor_power_dc_performance_max = values.dc.performance_max as u64;
-        self.processor_power_dc_boost_policy = values.dc.boost_policy as u64;
-        self.processor_power_dc_boost_mode = values.dc.boost_mode;
-    }
-
-    fn set_processor_power_boost_mode(
-        &mut self,
-        source: ProcessorPowerSource,
-        boost_mode: ProcessorBoostMode,
-    ) {
-        self.assign_processor_power_boost_mode(source, boost_mode);
-        if self.processor_power_link_ac_dc {
-            self.assign_processor_power_boost_mode(source.paired(), boost_mode);
-        }
-        self.active_power_plan_picker = None;
-        self.processor_power_dirty = true;
-    }
-
-    fn assign_processor_power_boost_mode(
-        &mut self,
-        source: ProcessorPowerSource,
-        boost_mode: ProcessorBoostMode,
-    ) {
-        match source {
-            ProcessorPowerSource::Ac => self.processor_power_ac_boost_mode = boost_mode,
-            ProcessorPowerSource::Dc => self.processor_power_dc_boost_mode = boost_mode,
-        }
-    }
-
-    fn set_processor_power_slider_value(&mut self, slider: ProcessorPowerSlider, value: u64) {
-        let value = value.min(100);
-        self.assign_processor_power_slider_value(slider, value);
-        if self.processor_power_link_ac_dc {
-            self.assign_processor_power_slider_value(slider.paired_power_source(), value);
-        }
-        self.processor_power_dirty = true;
-    }
-
-    fn assign_processor_power_slider_value(&mut self, slider: ProcessorPowerSlider, value: u64) {
-        match slider {
-            ProcessorPowerSlider::AcCoreParkingMin => {
-                self.processor_power_ac_core_parking_min = value;
-            }
-            ProcessorPowerSlider::AcPerformanceMin => {
-                self.processor_power_ac_performance_min = value;
-            }
-            ProcessorPowerSlider::AcPerformanceMax => {
-                self.processor_power_ac_performance_max = value;
-            }
-            ProcessorPowerSlider::AcBoostPolicy => {
-                self.processor_power_ac_boost_policy = value;
-            }
-            ProcessorPowerSlider::DcCoreParkingMin => {
-                self.processor_power_dc_core_parking_min = value;
-            }
-            ProcessorPowerSlider::DcPerformanceMin => {
-                self.processor_power_dc_performance_min = value;
-            }
-            ProcessorPowerSlider::DcPerformanceMax => {
-                self.processor_power_dc_performance_max = value;
-            }
-            ProcessorPowerSlider::DcBoostPolicy => {
-                self.processor_power_dc_boost_policy = value;
-            }
-        }
-    }
-
-    fn adaptive_engine_processor_policy_percent(
-        &self,
-        field: AdaptiveEngineProcessorPolicyField,
-    ) -> u32 {
-        let values = self
-            .settings
-            .adaptive_engine
-            .processor_policy_values
-            .normalized();
-        match field {
-            AdaptiveEngineProcessorPolicyField::CoreParkingMin => values.core_parking_min,
-            AdaptiveEngineProcessorPolicyField::PerformanceMin => values.performance_min,
-            AdaptiveEngineProcessorPolicyField::PerformanceMax => values.performance_max,
-            AdaptiveEngineProcessorPolicyField::BoostPolicy => values.boost_policy,
-        }
-    }
-
-    fn set_adaptive_engine_processor_policy_percent(
-        &mut self,
-        field: AdaptiveEngineProcessorPolicyField,
-        value: u64,
-    ) {
-        let mut values = self
-            .settings
-            .adaptive_engine
-            .processor_policy_values
-            .normalized();
-        let value = value.min(100) as u32;
-        match field {
-            AdaptiveEngineProcessorPolicyField::CoreParkingMin => values.core_parking_min = value,
-            AdaptiveEngineProcessorPolicyField::PerformanceMin => values.performance_min = value,
-            AdaptiveEngineProcessorPolicyField::PerformanceMax => values.performance_max = value,
-            AdaptiveEngineProcessorPolicyField::BoostPolicy => values.boost_policy = value,
-        }
-        self.settings.adaptive_engine.processor_policy_values = values.normalized();
-    }
-
-    fn sync_processor_power_slider_states(&self, window: &mut Window, cx: &mut Context<Self>) {
-        for (slider, value) in [
-            (
-                ProcessorPowerSlider::AcCoreParkingMin,
-                self.processor_power_ac_core_parking_min,
-            ),
-            (
-                ProcessorPowerSlider::AcPerformanceMin,
-                self.processor_power_ac_performance_min,
-            ),
-            (
-                ProcessorPowerSlider::AcPerformanceMax,
-                self.processor_power_ac_performance_max,
-            ),
-            (
-                ProcessorPowerSlider::AcBoostPolicy,
-                self.processor_power_ac_boost_policy,
-            ),
-            (
-                ProcessorPowerSlider::DcCoreParkingMin,
-                self.processor_power_dc_core_parking_min,
-            ),
-            (
-                ProcessorPowerSlider::DcPerformanceMin,
-                self.processor_power_dc_performance_min,
-            ),
-            (
-                ProcessorPowerSlider::DcPerformanceMax,
-                self.processor_power_dc_performance_max,
-            ),
-            (
-                ProcessorPowerSlider::DcBoostPolicy,
-                self.processor_power_dc_boost_policy,
-            ),
-        ] {
-            let input = processor_power_slider_input(&self.inputs, slider);
-            let value = value.min(100) as f32;
-            input.update(cx, |state, cx| {
-                if (state.value().end() - value).abs() > f32::EPSILON {
-                    state.set_value(value, window, cx);
-                }
-            });
-        }
-    }
-
-    fn set_cpu_threshold_slider_value(&mut self, slider: CpuThresholdSlider, value: u8) {
-        let value = value.min(100);
-        match slider {
-            CpuThresholdSlider::Lower(index) => {
-                if let Some(rule) = self.settings.by_cpu_load.rules.get_mut(index) {
-                    rule.threshold_percent = value;
-                }
-            }
-            CpuThresholdSlider::Upper(index) => {
-                if let Some(rule) = self.settings.by_cpu_load.rules.get_mut(index) {
-                    rule.upper_threshold_percent = Some(value);
-                }
-            }
-        }
-    }
-
-    fn sync_cpu_threshold_slider_states(&self, window: &mut Window, cx: &mut Context<Self>) {
-        for (index, rule) in self.settings.by_cpu_load.rules.iter().enumerate() {
-            self.sync_cpu_threshold_slider_state(
-                CpuThresholdSlider::Lower(index),
-                rule.threshold_percent,
-                window,
-                cx,
-            );
-            self.sync_cpu_threshold_slider_state(
-                CpuThresholdSlider::Upper(index),
-                rule.upper_threshold_percent.unwrap_or(100),
-                window,
-                cx,
-            );
-        }
-    }
-
-    fn sync_cpu_threshold_slider_state(
-        &self,
-        slider: CpuThresholdSlider,
-        value: u8,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let Some(input) = cpu_threshold_slider_input(&self.inputs, slider) else {
-            return;
-        };
-        let value = value.min(100) as f32;
-        input.update(cx, |state, cx| {
-            if (state.value().end() - value).abs() > f32::EPSILON {
-                state.set_value(value, window, cx);
-            }
-        });
-    }
-
-    fn set_activity_slider_value(&mut self, slider: ActivitySlider, value: u64) {
-        match slider {
-            ActivitySlider::IdleTimeout => {
-                self.settings.by_activity.idle_timeout_seconds = value.clamp(
-                    ACTIVITY_IDLE_TIMEOUT_MIN_SECONDS,
-                    ACTIVITY_IDLE_TIMEOUT_MAX_SECONDS,
-                );
-            }
-            ActivitySlider::CheckInterval => {
-                self.settings.general.check_interval_ms =
-                    snap_to_step(value, ACTIVITY_CHECK_INTERVAL_STEP_MS).clamp(
-                        ACTIVITY_CHECK_INTERVAL_MIN_MS,
-                        ACTIVITY_CHECK_INTERVAL_MAX_MS,
-                    );
-            }
-        }
-    }
-
-    fn sync_activity_slider_states(&self, window: &mut Window, cx: &mut Context<Self>) {
-        for (slider, input, value) in [
-            (
-                ActivitySlider::IdleTimeout,
-                self.inputs.activity_idle_timeout.clone(),
-                self.settings.by_activity.idle_timeout_seconds.clamp(
-                    ACTIVITY_IDLE_TIMEOUT_MIN_SECONDS,
-                    ACTIVITY_IDLE_TIMEOUT_MAX_SECONDS,
-                ),
-            ),
-            (
-                ActivitySlider::CheckInterval,
-                self.inputs.activity_check_interval.clone(),
-                self.settings.general.check_interval_ms.clamp(
-                    ACTIVITY_CHECK_INTERVAL_MIN_MS,
-                    ACTIVITY_CHECK_INTERVAL_MAX_MS,
-                ),
-            ),
-        ] {
-            let value = activity_slider_normalized_value(slider, value) as f32;
-            input.update(cx, |state, cx| {
-                if (state.value().end() - value).abs() > f32::EPSILON {
-                    state.set_value(value, window, cx);
-                }
-            });
-        }
-    }
-
-    fn refresh_processor_power_values(&mut self) {
-        let Some(plan) = self.processor_power_target_plan() else {
-            self.status_message = t!("processor_power.no_active_plan").to_string();
-            return;
-        };
-        if self.sync_processor_power_values_from_target_plan(true) {
-            self.status_message =
-                t!("processor_power.loaded_values", plan = plan.display_name()).to_string();
-        }
-    }
-
-    fn fill_processor_power_preset(&mut self, preset: ProcessorPowerPreset) {
-        let values = ProcessorPowerValues::for_preset(preset);
-        self.set_processor_power_values(ProcessorPowerAcDcValues::same(values));
-        self.processor_power_dirty = true;
-        self.status_message = t!(
-            "processor_power.loaded_preset",
-            preset = processor_power_preset_label(preset)
-        )
-        .to_string();
-    }
-
-    fn processor_power_matches_preset(&self, preset: ProcessorPowerPreset) -> bool {
-        let values = ProcessorPowerValues::for_preset(preset);
-        self.processor_power_values() == ProcessorPowerAcDcValues::same(values).normalized()
-    }
-
-    fn apply_processor_power_custom(&mut self) {
-        let Some(plan) = self.processor_power_target_plan() else {
-            self.status_message = t!("processor_power.no_active_plan").to_string();
-            return;
-        };
-
-        let values = self.processor_power_values();
-        self.set_processor_power_values(values);
-
-        match self
-            .power
-            .apply_processor_power_values(&plan.guid, values.normalized())
-        {
-            Ok(()) => {
-                self.processor_power_loaded_plan_guid = Some(plan.guid.clone());
-                self.processor_power_dirty = false;
-                self.status_message =
-                    t!("processor_power.applied_custom", plan = plan.display_name()).to_string();
-                self.refresh_active_plan();
-            }
-            Err(err) => self.status_message = err,
         }
     }
 
@@ -3154,1074 +2426,6 @@ impl WinderustApp {
         }
     }
 
-    fn rebuild_inputs(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        let settings = self.settings.clone();
-        let processor_power_values = self.processor_power_values();
-        self.editing_rule_title = None;
-        self.expanded_rule_cards.clear();
-        self.inputs = UiInputs::new(window, cx, &settings, processor_power_values);
-        self.rebuild_rule_title_input_subscriptions(window, cx);
-        self.subscribe_to_dashboard_search_input(window, cx);
-        self.subscribe_to_processor_power_sliders(window, cx);
-        self.rebuild_cpu_threshold_slider_subscriptions(window, cx);
-        self.subscribe_to_activity_sliders(window, cx);
-    }
-
-    fn rule_title_input_count(&self) -> usize {
-        self.inputs.by_time_rule_names.len() + self.inputs.by_cpu_load_rule_names.len()
-    }
-
-    fn ensure_rule_title_input_subscriptions(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if self._rule_title_input_subscriptions.len() != self.rule_title_input_count() {
-            self.rebuild_rule_title_input_subscriptions(window, cx);
-        }
-    }
-
-    fn rebuild_rule_title_input_subscriptions(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let mut inputs = Vec::new();
-        inputs.extend(
-            self.inputs
-                .by_time_rule_names
-                .iter()
-                .cloned()
-                .enumerate()
-                .map(|(index, input)| (input, RuleTitleTarget::ByTime(index))),
-        );
-        inputs.extend(
-            self.inputs
-                .by_cpu_load_rule_names
-                .iter()
-                .cloned()
-                .enumerate()
-                .map(|(index, input)| (input, RuleTitleTarget::ByCpuLoad(index))),
-        );
-
-        self._rule_title_input_subscriptions.clear();
-        for (input, target) in inputs {
-            self.subscribe_to_rule_title_input(input, target, window, cx);
-        }
-    }
-
-    fn subscribe_to_rule_title_input(
-        &mut self,
-        input: Entity<InputState>,
-        target: RuleTitleTarget,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self._rule_title_input_subscriptions.push(cx.subscribe_in(
-            &input,
-            window,
-            move |app, _, event: &InputEvent, _, cx| {
-                app.handle_rule_title_input_event(target, event, cx);
-            },
-        ));
-    }
-
-    fn handle_rule_title_input_event(
-        &mut self,
-        target: RuleTitleTarget,
-        event: &InputEvent,
-        cx: &mut Context<Self>,
-    ) {
-        if matches!(event, InputEvent::PressEnter { .. } | InputEvent::Blur) {
-            self.finish_rule_title_edit(target, cx);
-        }
-    }
-
-    fn subscribe_to_numeric_input(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        self._numeric_input_subscription = Some(cx.subscribe_in(
-            &self.inputs.numeric_value,
-            window,
-            move |app, _, event: &InputEvent, _, cx| {
-                app.handle_numeric_input_event(event, cx);
-            },
-        ));
-    }
-
-    fn subscribe_to_dashboard_search_input(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        self._dashboard_search_subscription = Some(cx.subscribe_in(
-            &self.inputs.dashboard_search,
-            window,
-            move |_, _, _: &InputEvent, _, cx| {
-                cx.notify();
-            },
-        ));
-    }
-
-    fn subscribe_to_processor_power_sliders(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self._processor_power_slider_subscriptions.clear();
-        for slider in [
-            ProcessorPowerSlider::AcCoreParkingMin,
-            ProcessorPowerSlider::AcPerformanceMin,
-            ProcessorPowerSlider::AcPerformanceMax,
-            ProcessorPowerSlider::AcBoostPolicy,
-            ProcessorPowerSlider::DcCoreParkingMin,
-            ProcessorPowerSlider::DcPerformanceMin,
-            ProcessorPowerSlider::DcPerformanceMax,
-            ProcessorPowerSlider::DcBoostPolicy,
-        ] {
-            let input = processor_power_slider_input(&self.inputs, slider);
-            self._processor_power_slider_subscriptions
-                .push(
-                    cx.subscribe_in(&input, window, move |app, _, event, _, cx| {
-                        app.handle_processor_power_slider_event(slider, event, cx);
-                    }),
-                );
-        }
-    }
-
-    fn handle_processor_power_slider_event(
-        &mut self,
-        slider: ProcessorPowerSlider,
-        event: &SliderEvent,
-        cx: &mut Context<Self>,
-    ) {
-        let SliderEvent::Change(value) = event;
-        self.set_processor_power_slider_value(slider, value.end().round() as u64);
-        cx.notify();
-    }
-
-    fn cpu_threshold_slider_input_count(&self) -> usize {
-        self.inputs.cpu_rule_thresholds.len() + self.inputs.cpu_rule_upper_thresholds.len()
-    }
-
-    fn ensure_cpu_threshold_slider_subscriptions(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if self._cpu_threshold_slider_subscriptions.len() != self.cpu_threshold_slider_input_count()
-        {
-            self.rebuild_cpu_threshold_slider_subscriptions(window, cx);
-        }
-    }
-
-    fn rebuild_cpu_threshold_slider_subscriptions(
-        &mut self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let mut inputs = Vec::new();
-        inputs.extend(
-            self.inputs
-                .cpu_rule_thresholds
-                .iter()
-                .cloned()
-                .enumerate()
-                .map(|(index, input)| (input, CpuThresholdSlider::Lower(index))),
-        );
-        inputs.extend(
-            self.inputs
-                .cpu_rule_upper_thresholds
-                .iter()
-                .cloned()
-                .enumerate()
-                .map(|(index, input)| (input, CpuThresholdSlider::Upper(index))),
-        );
-
-        self._cpu_threshold_slider_subscriptions.clear();
-        for (input, slider) in inputs {
-            self._cpu_threshold_slider_subscriptions
-                .push(cx.subscribe_in(
-                    &input,
-                    window,
-                    move |app, _, event: &SliderEvent, _, cx| {
-                        app.handle_cpu_threshold_slider_event(slider, event, cx);
-                    },
-                ));
-        }
-    }
-
-    fn handle_cpu_threshold_slider_event(
-        &mut self,
-        slider: CpuThresholdSlider,
-        event: &SliderEvent,
-        cx: &mut Context<Self>,
-    ) {
-        let SliderEvent::Change(value) = event;
-        let value = value.end().round().clamp(0.0, 100.0) as u8;
-        self.set_cpu_threshold_slider_value(slider, value);
-        cx.notify();
-    }
-
-    fn subscribe_to_activity_sliders(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        self._activity_slider_subscriptions.clear();
-        for (slider, input) in [
-            (
-                ActivitySlider::IdleTimeout,
-                self.inputs.activity_idle_timeout.clone(),
-            ),
-            (
-                ActivitySlider::CheckInterval,
-                self.inputs.activity_check_interval.clone(),
-            ),
-        ] {
-            self._activity_slider_subscriptions.push(cx.subscribe_in(
-                &input,
-                window,
-                move |app, _, event: &SliderEvent, _, cx| {
-                    app.handle_activity_slider_event(slider, event, cx);
-                },
-            ));
-        }
-    }
-
-    fn handle_activity_slider_event(
-        &mut self,
-        slider: ActivitySlider,
-        event: &SliderEvent,
-        cx: &mut Context<Self>,
-    ) {
-        let SliderEvent::Change(value) = event;
-        self.set_activity_slider_value(slider, value.end().round() as u64);
-        cx.notify();
-    }
-
-    fn handle_numeric_input_event(&mut self, event: &InputEvent, cx: &mut Context<Self>) {
-        if matches!(event, InputEvent::PressEnter { .. } | InputEvent::Blur) {
-            self.finish_numeric_edit(cx);
-        }
-    }
-
-    fn rule_title_input(&self, target: RuleTitleTarget) -> Option<Entity<InputState>> {
-        match target {
-            RuleTitleTarget::ByTime(index) => self.inputs.by_time_rule_names.get(index),
-            RuleTitleTarget::ByCpuLoad(index) => self.inputs.by_cpu_load_rule_names.get(index),
-        }
-        .cloned()
-    }
-
-    fn begin_rule_title_edit(
-        &mut self,
-        target: RuleTitleTarget,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.editing_rule_title = Some(target);
-        if let Some(input) = self.rule_title_input(target) {
-            input.read(cx).focus_handle(cx).focus(window);
-        }
-        cx.notify();
-    }
-
-    fn begin_numeric_edit(
-        &mut self,
-        field: NumericField,
-        value: String,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.editing_numeric = Some(field);
-        clear_input_to(&self.inputs.numeric_value, &value, window, cx);
-        self.inputs
-            .numeric_value
-            .read(cx)
-            .focus_handle(cx)
-            .focus(window);
-        cx.notify();
-    }
-
-    fn finish_numeric_edit(&mut self, cx: &mut Context<Self>) {
-        let Some(field) = self.editing_numeric.take() else {
-            return;
-        };
-        let value = self.inputs.numeric_value.read(cx).value().to_string();
-        self.apply_numeric_input(field, &value);
-        cx.notify();
-    }
-
-    fn apply_numeric_input(&mut self, field: NumericField, value: &str) {
-        let value = value.trim().replace(',', "");
-        match field {
-            NumericField::ActivityIdleTimeout => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    ACTIVITY_IDLE_TIMEOUT_MIN_SECONDS,
-                    ACTIVITY_IDLE_TIMEOUT_MAX_SECONDS,
-                ) {
-                    self.set_activity_slider_value(ActivitySlider::IdleTimeout, value);
-                }
-            }
-            NumericField::GeneralCheckInterval => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    ACTIVITY_CHECK_INTERVAL_MIN_MS,
-                    ACTIVITY_CHECK_INTERVAL_MAX_MS,
-                ) {
-                    self.set_activity_slider_value(ActivitySlider::CheckInterval, value);
-                }
-            }
-            NumericField::ExecutionFailureSuppressionThreshold => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    u64::from(MIN_EXECUTION_FAILURE_SUPPRESSION_THRESHOLD),
-                    u64::from(MAX_EXECUTION_FAILURE_SUPPRESSION_THRESHOLD),
-                ) {
-                    self.settings
-                        .advanced
-                        .execution_failure_suppression_threshold = value as u8;
-                }
-            }
-            NumericField::BackgroundCpuRestrictionPercent => {
-                if let Some(value) = parse_u64_input(&value, 1, 100) {
-                    self.settings.background_cpu_restriction.percent = value as u8;
-                }
-            }
-            NumericField::MemoryTrimCheckIntervalMinutes => {
-                if let Some(value) = parse_u64_input(&value, 1, 1440) {
-                    self.settings.memory_trim.check_interval_minutes = value;
-                }
-            }
-            NumericField::MemoryTrimMemoryLoadThreshold => {
-                if let Some(value) = parse_u64_input(&value, 1, 100) {
-                    self.settings
-                        .memory_trim
-                        .system_memory_load_threshold_percent = value as u8;
-                }
-            }
-            NumericField::MemoryTrimWorkingSetThreshold => {
-                if let Some(value) = parse_u64_input(&value, 1, 1_048_576) {
-                    self.settings.memory_trim.process_working_set_threshold_mb = value;
-                }
-            }
-            NumericField::MemoryTrimCpuIdleThreshold => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.settings.memory_trim.process_cpu_idle_threshold_percent = value as u8;
-                }
-            }
-            NumericField::MemoryTrimIdleSeconds => {
-                if let Some(value) = parse_u64_input(&value, 1, 86_400) {
-                    self.settings.memory_trim.process_idle_seconds = value;
-                }
-            }
-            NumericField::MemoryTrimCooldownSeconds => {
-                if let Some(value) = parse_u64_input(&value, 1, 86_400) {
-                    self.settings.memory_trim.trim_cooldown_seconds = value;
-                }
-            }
-            NumericField::MemoryTrimPurgeFreeRamThreshold => {
-                if let Some(value) = parse_u64_input(&value, 0, 1_048_576) {
-                    self.settings.memory_trim.purge_free_ram_threshold_mb = value;
-                }
-            }
-            NumericField::SuspensionBackgroundDelay => {
-                if let Some(value) = parse_u64_input(&value, 1, 86_400) {
-                    self.settings.app_suspension.background_delay_seconds = value;
-                }
-            }
-            NumericField::SuspensionThawInterval => {
-                if let Some(value) = parse_u64_input(&value, 1, 86_400) {
-                    self.settings.app_suspension.temporary_thaw_interval_seconds = value;
-                }
-            }
-            NumericField::SuspensionThawDuration => {
-                if let Some(value) = parse_u64_input(&value, 1, 3_600) {
-                    self.settings.app_suspension.temporary_thaw_duration_seconds = value;
-                }
-            }
-            NumericField::SuspensionAudioRefreeze => {
-                if let Some(value) = parse_u64_input(&value, 1, 3_600) {
-                    self.settings.app_suspension.audio_wake_duration_seconds = value;
-                }
-            }
-            NumericField::SuspensionNetworkRefreeze => {
-                if let Some(value) = parse_u64_input(&value, 1, 3_600) {
-                    self.settings.app_suspension.network_wake_duration_seconds = value;
-                }
-            }
-            NumericField::WorkloadEngineThreshold => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_THRESHOLD_MIN_PERCENT,
-                    WORKLOAD_ENGINE_THRESHOLD_MAX_PERCENT,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_threshold_percent = value as u8;
-                }
-            }
-            NumericField::WorkloadEngineRestoreThreshold => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_THRESHOLD_MIN_PERCENT,
-                    WORKLOAD_ENGINE_THRESHOLD_MAX_PERCENT,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_restore_threshold_percent = value as u8;
-                }
-            }
-            NumericField::WorkloadEngineTotalThreshold => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_THRESHOLD_MIN_PERCENT,
-                    WORKLOAD_ENGINE_THRESHOLD_MAX_PERCENT,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_total_threshold_percent = value as u8;
-                }
-            }
-            NumericField::WorkloadEngineCpuPercent => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_THRESHOLD_MIN_PERCENT,
-                    WORKLOAD_ENGINE_THRESHOLD_MAX_PERCENT,
-                ) {
-                    self.settings.workload_engine.workload_engine_cpu_percent = value as u8;
-                }
-            }
-            NumericField::WorkloadEngineSustain => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_SECONDS_MIN,
-                    WORKLOAD_ENGINE_SECONDS_MAX,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_sustain_seconds = value;
-                }
-            }
-            NumericField::WorkloadEngineMinimumRestraint => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_SECONDS_MIN,
-                    WORKLOAD_ENGINE_SECONDS_MAX,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_minimum_restraint_seconds = value;
-                }
-            }
-            NumericField::WorkloadEngineCooldown => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_SECONDS_MIN,
-                    WORKLOAD_ENGINE_SECONDS_MAX,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_cooldown_seconds = value;
-                }
-            }
-            NumericField::WorkloadEngineMaxTargetedProcesses => {
-                if let Some(value) = parse_u64_input(
-                    &value,
-                    WORKLOAD_ENGINE_TARGET_LIMIT_MIN,
-                    WORKLOAD_ENGINE_TARGET_LIMIT_MAX,
-                ) {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_max_targeted_processes = value as u8;
-                }
-            }
-            NumericField::ProcessorAcCoreParkingMin => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::AcCoreParkingMin,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorAcPerformanceMin => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::AcPerformanceMin,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorAcPerformanceMax => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::AcPerformanceMax,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorAcBoostPolicy => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::AcBoostPolicy,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorDcCoreParkingMin => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::DcCoreParkingMin,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorDcPerformanceMin => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::DcPerformanceMin,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorDcPerformanceMax => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::DcPerformanceMax,
-                        value,
-                    );
-                }
-            }
-            NumericField::ProcessorDcBoostPolicy => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_processor_power_slider_value(
-                        ProcessorPowerSlider::DcBoostPolicy,
-                        value,
-                    );
-                }
-            }
-            NumericField::AdaptiveEngineProcessorPolicy(field) => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_adaptive_engine_processor_policy_percent(field, value);
-                }
-            }
-            NumericField::CpuThreshold(index) => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_cpu_threshold_slider_value(
-                        CpuThresholdSlider::Lower(index),
-                        value as u8,
-                    );
-                }
-            }
-            NumericField::CpuUpperThreshold(index) => {
-                if let Some(value) = parse_u64_input(&value, 0, 100) {
-                    self.set_cpu_threshold_slider_value(
-                        CpuThresholdSlider::Upper(index),
-                        value as u8,
-                    );
-                }
-            }
-            NumericField::CpuDuration(index) => {
-                if let (Some(rule), Some(value)) = (
-                    self.settings.by_cpu_load.rules.get_mut(index),
-                    parse_u64_input(&value, 0, 86_400),
-                ) {
-                    rule.duration_seconds = value;
-                }
-            }
-            NumericField::CoreLimiterThreshold(index) => {
-                if let (Some(rule), Some(value)) = (
-                    self.settings.core_limiter.rules.get_mut(index),
-                    parse_u64_input(&value, 1, 100),
-                ) {
-                    rule.threshold_percent = value as u8;
-                }
-            }
-            NumericField::CoreLimiterSustain(index) => {
-                if let (Some(rule), Some(value)) = (
-                    self.settings.core_limiter.rules.get_mut(index),
-                    parse_u64_input(&value, 1, 86_400),
-                ) {
-                    rule.sustain_seconds = value;
-                }
-            }
-            NumericField::CoreLimiterCooldown(index) => {
-                if let (Some(rule), Some(value)) = (
-                    self.settings.core_limiter.rules.get_mut(index),
-                    parse_u64_input(&value, 1, 86_400),
-                ) {
-                    rule.cooldown_seconds = value;
-                }
-            }
-            NumericField::CoreLimiterMaxProcessors(index) => {
-                if let (Some(rule), Some(value)) = (
-                    self.settings.core_limiter.rules.get_mut(index),
-                    parse_u64_input(&value, 1, max_logical_processor_count() as u64),
-                ) {
-                    rule.max_logical_processors = value as u8;
-                }
-            }
-            NumericField::TimerResolutionRule(index) => {
-                let minimum_100ns = self
-                    .timer_resolution_status
-                    .minimum_100ns
-                    .unwrap_or((TIMER_RESOLUTION_INPUT_MIN_MS * 10_000.0).round() as u32);
-                let maximum_100ns = self
-                    .timer_resolution_status
-                    .maximum_100ns
-                    .unwrap_or((TIMER_RESOLUTION_INPUT_MAX_MS * 10_000.0).round() as u32);
-                if let (Some(rule), Some(value)) = (
-                    self.settings.timer_resolution.rules.get_mut(index),
-                    parse_timer_resolution_input_100ns(&value, minimum_100ns, maximum_100ns),
-                ) {
-                    rule.desired_100ns = value;
-                }
-            }
-            NumericField::NetworkThreshold(field) => {
-                let Ok(value) = value.parse::<f64>() else {
-                    return;
-                };
-                let Some(rule) = self.threshold_rule_mut(field) else {
-                    return;
-                };
-                let (bytes, unit) = match field {
-                    ThresholdField::Download(_) => (
-                        &mut rule.network_download_threshold_bytes,
-                        rule.network_download_threshold_unit,
-                    ),
-                    ThresholdField::Upload(_) => (
-                        &mut rule.network_upload_threshold_bytes,
-                        rule.network_upload_threshold_unit,
-                    ),
-                };
-                *bytes = unit
-                    .threshold_bytes_from_value(value.max(0.0))
-                    .min(MAX_NETWORK_THRESHOLD_BYTES);
-            }
-        }
-    }
-
-    fn finish_rule_title_edit(&mut self, target: RuleTitleTarget, cx: &mut Context<Self>) {
-        self.sync_input_values(cx);
-        if self.editing_rule_title == Some(target) {
-            self.editing_rule_title = None;
-        }
-        cx.notify();
-    }
-
-    fn is_rule_card_collapsed(&self, target: &RuleCardTarget) -> bool {
-        !self.expanded_rule_cards.contains(target)
-    }
-
-    fn toggle_rule_card(&mut self, target: RuleCardTarget, cx: &mut Context<Self>) {
-        let motion_id = rule_card_body_motion_id(&target);
-        let expanded = if self.expanded_rule_cards.remove(&target) {
-            false
-        } else {
-            self.expanded_rule_cards.insert(target);
-            true
-        };
-        begin_expandable_motion(motion_id, expanded);
-        cx.notify();
-    }
-
-    fn is_setting_group_collapsed(&self, target: SettingGroupTarget) -> bool {
-        !self.expanded_setting_groups.contains(&target)
-    }
-
-    fn toggle_setting_group(&mut self, target: SettingGroupTarget, cx: &mut Context<Self>) {
-        self.set_setting_group_expanded(target, self.is_setting_group_collapsed(target));
-        cx.notify();
-    }
-
-    fn set_setting_group_expanded(&mut self, target: SettingGroupTarget, expanded: bool) {
-        let changed = if expanded {
-            self.expanded_setting_groups.insert(target)
-        } else {
-            self.expanded_setting_groups.remove(&target)
-        };
-
-        if changed {
-            begin_expandable_motion(format!("setting-group-{target:?}"), expanded);
-        }
-    }
-
-    fn is_process_list_group_collapsed(&self, process_name: &str) -> bool {
-        !self
-            .expanded_process_list_groups
-            .contains(&process_list_group_key(process_name))
-    }
-
-    fn toggle_process_list_group(&mut self, process_name: String, cx: &mut Context<Self>) {
-        let key = process_list_group_key(&process_name);
-        let expanded = if self.expanded_process_list_groups.remove(&key) {
-            false
-        } else {
-            self.expanded_process_list_groups.insert(key.clone());
-            true
-        };
-        begin_expandable_motion(format!("process-list-group-{key}"), expanded);
-        cx.notify();
-    }
-
-    fn set_process_list_column_visible(
-        &mut self,
-        column: ProcessListColumn,
-        visible: bool,
-        cx: &mut Context<Self>,
-    ) {
-        let changed = if visible {
-            self.hidden_process_list_columns.remove(&column)
-        } else {
-            self.hidden_process_list_columns.insert(column)
-        };
-
-        let sort_changed =
-            !visible && self.process_list_sort.column == ProcessListSortColumn::Column(column);
-        if sort_changed {
-            self.process_list_sort = ProcessListSort::default();
-        }
-
-        if changed || sort_changed {
-            cx.notify();
-        }
-    }
-
-    fn toggle_process_list_sort(&mut self, column: ProcessListSortColumn, cx: &mut Context<Self>) {
-        self.process_list_sort = self.process_list_sort.toggled_for(column);
-        cx.notify();
-    }
-
-    fn finish_process_list_edit(&mut self, cx: &mut Context<Self>) {
-        self.active_power_plan_picker = None;
-        cx.notify();
-    }
-
-    fn set_process_list_foreground_power_plan(
-        &mut self,
-        process_name: String,
-        power_plan_guid: Option<String>,
-        cx: &mut Context<Self>,
-    ) {
-        set_foreground_power_plan_override(
-            &mut self.settings.by_foreground,
-            &process_name,
-            power_plan_guid,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_running_power_plan(
-        &mut self,
-        process_name: String,
-        power_plan_guid: Option<String>,
-        cx: &mut Context<Self>,
-    ) {
-        set_by_running_app_power_plan_override(
-            &mut self.settings.by_running_app,
-            &process_name,
-            power_plan_guid,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_background_efficiency(
-        &mut self,
-        process_name: String,
-        included: bool,
-        cx: &mut Context<Self>,
-    ) {
-        set_background_efficiency_custom_rule(
-            &mut self.settings.background_efficiency,
-            &process_name,
-            !included,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_background_cpu_restriction(
-        &mut self,
-        process_name: String,
-        included: bool,
-        cx: &mut Context<Self>,
-    ) {
-        set_process_exclusion(
-            &mut self.settings.background_cpu_restriction.exclusions,
-            &process_name,
-            !included,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_core_limiter(
-        &mut self,
-        process_name: String,
-        max_logical_processors: Option<u8>,
-        cx: &mut Context<Self>,
-    ) {
-        set_core_limiter_override(
-            &mut self.settings.core_limiter,
-            &process_name,
-            max_logical_processors,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_gpu_priority_included(
-        &mut self,
-        process_name: String,
-        included: bool,
-        cx: &mut Context<Self>,
-    ) {
-        set_process_exclusion(
-            &mut self.settings.gpu_priority.exclusions,
-            &process_name,
-            !included,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_memory_trim(
-        &mut self,
-        process_name: String,
-        included: bool,
-        cx: &mut Context<Self>,
-    ) {
-        set_process_exclusion(
-            &mut self.settings.memory_trim.exclusions,
-            &process_name,
-            !included,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_app_suspension(
-        &mut self,
-        process_name: String,
-        included: bool,
-        cx: &mut Context<Self>,
-    ) {
-        set_app_suspension_override(&mut self.settings.app_suspension, &process_name, included);
-        self.finish_process_list_edit(cx);
-    }
-
-    fn set_process_list_timer_resolution(
-        &mut self,
-        process_name: String,
-        desired_100ns: Option<u32>,
-        cx: &mut Context<Self>,
-    ) {
-        set_timer_resolution_override(
-            &mut self.settings.timer_resolution,
-            &process_name,
-            desired_100ns,
-        );
-        self.finish_process_list_edit(cx);
-    }
-
-    fn apply_process_priority_once(
-        &mut self,
-        target: Result<ProcessActionTarget, ProcessActionTargetError>,
-        process_name: &str,
-        priority: ProcessPrioritySetting,
-        cx: &mut Context<Self>,
-    ) {
-        self.status_message = match target
-            .map_err(|error| error.to_string())
-            .and_then(|target| process_priority::apply_once(&target, priority))
-        {
-            Ok(priority) => t!(
-                "process_list.applied_once",
-                name = process_name,
-                priority = priority
-            )
-            .to_string(),
-            Err(error) => t!(
-                "process_list.apply_once_failed",
-                name = process_name,
-                error = error
-            )
-            .to_string(),
-        };
-        cx.notify();
-    }
-
-    fn save_process_priority_rule(
-        &mut self,
-        process_name: &str,
-        priority: ProcessPrioritySetting,
-        cx: &mut Context<Self>,
-    ) {
-        set_process_priority_rule(&mut self.settings.process_priority, process_name, priority);
-        if self.save_settings() {
-            let key = if self.settings.process_priority.enabled {
-                "process_list.saved_priority_rule"
-            } else {
-                "process_list.saved_priority_rule_disabled"
-            };
-            self.status_message = t!(
-                key,
-                name = process_name,
-                priority = process_priority_setting_label(priority)
-            )
-            .to_string();
-        }
-        cx.notify();
-    }
-
-    fn apply_memory_priority_once(
-        &mut self,
-        target: Result<ProcessActionTarget, ProcessActionTargetError>,
-        process_name: &str,
-        priority: ProcessMemoryPrioritySetting,
-        cx: &mut Context<Self>,
-    ) {
-        self.status_message = match target
-            .map_err(|error| error.to_string())
-            .and_then(|target| memory_priority::apply_once(&target, priority))
-        {
-            Ok(priority) => t!(
-                "process_list.applied_memory_once",
-                name = process_name,
-                priority = priority
-            )
-            .to_string(),
-            Err(error) => t!(
-                "process_list.apply_memory_once_failed",
-                name = process_name,
-                error = error
-            )
-            .to_string(),
-        };
-        cx.notify();
-    }
-
-    fn save_memory_priority_rule(
-        &mut self,
-        process_name: &str,
-        priority: ProcessMemoryPrioritySetting,
-        cx: &mut Context<Self>,
-    ) {
-        set_memory_priority_rule(&mut self.settings.memory_priority, process_name, priority);
-        if self.save_settings() {
-            let key = if self.settings.memory_priority.enabled {
-                "process_list.saved_memory_rule"
-            } else {
-                "process_list.saved_memory_rule_disabled"
-            };
-            self.status_message = t!(
-                key,
-                name = process_name,
-                priority = process_memory_priority_setting_label(priority)
-            )
-            .to_string();
-        }
-        cx.notify();
-    }
-
-    fn dropdown_placement(
-        &self,
-        id: &str,
-        full_list_height: Pixels,
-        window: &Window,
-    ) -> DropdownPlacement {
-        let margin = px(DROPDOWN_VIEWPORT_MARGIN);
-        let offset = px(DROPDOWN_MENU_OFFSET);
-        let fallback_max_height =
-            (window.viewport_size().height - offset - margin).max(Pixels::ZERO);
-        let Some(bounds) = self.dropdown_anchor_bounds.borrow().get(id).copied() else {
-            return DropdownPlacement {
-                open_up: false,
-                max_height: fallback_max_height,
-            };
-        };
-
-        let below =
-            (window.viewport_size().height - bounds.top() - offset - margin).max(Pixels::ZERO);
-        let above = (bounds.bottom() - offset - margin).max(Pixels::ZERO);
-        let open_up = full_list_height > below && above > below;
-        let available_height = if open_up { above } else { below };
-
-        DropdownPlacement {
-            open_up,
-            max_height: available_height,
-        }
-    }
-
-    #[expect(
-        clippy::too_many_arguments,
-        reason = "dropdown metadata plus the options builder is clearer at each UI call site"
-    )]
-    fn render_dropdown_select(
-        &self,
-        id: impl Into<String>,
-        selected_label: impl Into<SharedString>,
-        enabled: bool,
-        width: DropdownSelectWidth,
-        option_count: usize,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-        build_options: impl FnOnce(Pixels, &mut Context<Self>) -> Scrollable<gpui::Div>,
-    ) -> AnyElement {
-        let id = id.into();
-        let is_open = enabled && self.active_power_plan_picker.as_deref() == Some(id.as_str());
-        let placement = self.dropdown_placement(&id, dropdown_list_height(option_count), window);
-        let options = build_options(placement.max_height, cx);
-        let phase = dropdown_popup_phase(id.as_str(), is_open, cx);
-        let control_id = SharedString::from(format!("{id}-control"));
-        let toggle_id = id.clone();
-
-        dropdown_select_container(width)
-            .child(
-                dropdown_select_control(control_id, selected_label, enabled, is_open, phase, cx)
-                    .when(enabled, |control| {
-                        control.on_click(cx.listener(move |app, _: &gpui::ClickEvent, _, cx| {
-                            app.active_power_plan_picker =
-                                (app.active_power_plan_picker.as_deref()
-                                    != Some(toggle_id.as_str()))
-                                .then_some(toggle_id.clone());
-                            cx.stop_propagation();
-                            cx.notify();
-                        }))
-                    }),
-            )
-            .child(dropdown_anchor_sensor(
-                id.clone(),
-                Rc::clone(&self.dropdown_anchor_bounds),
-            ))
-            .child(dropdown_popup_or_empty(
-                SharedString::from(id),
-                phase,
-                placement,
-                options,
-                cx,
-            ))
-            .into_any_element()
-    }
-
-    fn sync_input_values(&mut self, cx: &mut Context<Self>) {
-        for (rule, input) in self
-            .settings
-            .by_cpu_load
-            .rules
-            .iter_mut()
-            .zip(&self.inputs.by_cpu_load_rule_names)
-        {
-            rule.name = input.read(cx).value().to_string();
-        }
-        for (index, rule) in self.settings.by_time.rules.iter_mut().enumerate() {
-            if let Some(input) = self.inputs.by_time_rule_names.get(index) {
-                rule.name = input.read(cx).value().to_string();
-            }
-            if let Some(input) = self.inputs.schedule_start_times.get(index) {
-                rule.start_time = input.read(cx).value().to_string();
-            }
-            if let Some(input) = self.inputs.schedule_end_times.get(index) {
-                rule.end_time = input.read(cx).value().to_string();
-            }
-        }
-        for (index, rule) in self.settings.by_foreground.rules.iter_mut().enumerate() {
-            if let Some(input) = self.inputs.foreground_rule_names.get(index) {
-                rule.name = input.read(cx).value().to_string();
-            }
-            if let Some(input) = self.inputs.foreground_rule_processes.get(index) {
-                rule.process_name = input.read(cx).value().to_string();
-            }
-        }
-    }
-
     fn background_settings(&self) -> Settings {
         self.runtime_settings()
     }
@@ -4915,763 +3119,6 @@ mod tests {
             &current,
             &saved
         ));
-    }
-
-    #[test]
-    fn process_list_column_layout_fits_headers_and_values() {
-        let settings = Settings::default();
-        let processes = vec![
-            ProcessInfo {
-                id: 1234,
-                parent_id: None,
-                name: "editor.exe".to_owned(),
-            },
-            ProcessInfo {
-                id: 12345,
-                parent_id: None,
-                name: "worker.exe".to_owned(),
-            },
-        ];
-        let groups = process_list_groups(&processes);
-        let summaries = groups
-            .iter()
-            .map(|_| default_process_policy_summary())
-            .collect::<Vec<_>>();
-
-        let layout = process_list_column_layout(&settings, &groups, &summaries);
-
-        assert!(layout.column_width(ProcessListColumn::Pid) < PROCESS_LIST_PID_MAX_WIDTH);
-        assert!(layout.column_width(ProcessListColumn::MemoryTrim) < 140.0);
-        assert!(
-            layout.column_width(ProcessListColumn::PowerPlanForeground)
-                >= process_list_estimated_cell_width(
-                    &process_list_column_label(ProcessListColumn::PowerPlanForeground, &settings),
-                    PROCESS_LIST_TEXT_CELL_HORIZONTAL_PADDING,
-                )
-        );
-    }
-
-    #[test]
-    fn process_icon_cache_drops_stale_paths() {
-        let kept_path = PathBuf::from("C:\\Apps\\kept.exe");
-        let stale_path = PathBuf::from("C:\\Apps\\stale.exe");
-        let mut cache = HashMap::from([(kept_path.clone(), None), (stale_path.clone(), None)]);
-        let candidates = vec![ProcessCandidate {
-            name: "kept.exe".to_owned(),
-            image_path: Some(kept_path.clone()),
-            icon: None,
-        }];
-
-        WinderustApp::retain_current_process_icons(&mut cache, &candidates);
-
-        assert!(cache.contains_key(&kept_path));
-        assert!(!cache.contains_key(&stale_path));
-    }
-
-    #[test]
-    fn process_list_sort_orders_groups_by_name_direction() {
-        let processes = vec![
-            ProcessInfo {
-                id: 1,
-                parent_id: None,
-                name: "editor.exe".to_owned(),
-            },
-            ProcessInfo {
-                id: 2,
-                parent_id: None,
-                name: "worker.exe".to_owned(),
-            },
-        ];
-        let groups = process_list_groups(&processes);
-        let summaries = groups
-            .iter()
-            .map(|_| default_process_policy_summary())
-            .collect::<Vec<_>>();
-        let rows = process_list_sorted_rows(
-            groups,
-            summaries,
-            ProcessListSort {
-                column: ProcessListSortColumn::ProcessName,
-                direction: ProcessListSortDirection::Descending,
-            },
-        );
-
-        assert_eq!(rows[0].0.display_name, "worker.exe");
-        assert_eq!(rows[1].0.display_name, "editor.exe");
-    }
-
-    #[test]
-    fn process_list_text_sort_cmp_matches_ascii_lowercase_sorting() {
-        for (left, right) in [
-            ("Alpha.exe", "alpha.exe"),
-            ("worker.exe", "Editor.exe"),
-            ("z.exe", "é.exe"),
-        ] {
-            let expected = left
-                .to_ascii_lowercase()
-                .cmp(&right.to_ascii_lowercase())
-                .then_with(|| left.cmp(right));
-            assert_eq!(process_list_text_sort_cmp(left, right), expected);
-        }
-    }
-
-    #[test]
-    fn process_list_sort_orders_groups_and_children_by_pid() {
-        let processes = vec![
-            ProcessInfo {
-                id: 30,
-                parent_id: None,
-                name: "editor.exe".to_owned(),
-            },
-            ProcessInfo {
-                id: 10,
-                parent_id: None,
-                name: "worker.exe".to_owned(),
-            },
-            ProcessInfo {
-                id: 20,
-                parent_id: None,
-                name: "editor.exe".to_owned(),
-            },
-        ];
-        let sort = ProcessListSort {
-            column: ProcessListSortColumn::Column(ProcessListColumn::Pid),
-            direction: ProcessListSortDirection::Ascending,
-        };
-        let mut groups = process_list_groups(&processes);
-        for group in &mut groups {
-            process_list_sort_group_processes(group, sort);
-        }
-        let summaries = groups
-            .iter()
-            .map(|_| default_process_policy_summary())
-            .collect::<Vec<_>>();
-        let rows = process_list_sorted_rows(groups, summaries, sort);
-
-        assert_eq!(rows[0].0.display_name, "worker.exe");
-        assert_eq!(rows[1].0.display_name, "editor.exe");
-        assert_eq!(rows[1].0.processes[0].id, 20);
-        assert_eq!(rows[1].0.processes[1].id, 30);
-
-        let sort = ProcessListSort {
-            column: ProcessListSortColumn::Column(ProcessListColumn::Pid),
-            direction: ProcessListSortDirection::Descending,
-        };
-        let mut groups = process_list_groups(&processes);
-        for group in &mut groups {
-            process_list_sort_group_processes(group, sort);
-        }
-        let summaries = groups
-            .iter()
-            .map(|_| default_process_policy_summary())
-            .collect::<Vec<_>>();
-        let rows = process_list_sorted_rows(groups, summaries, sort);
-
-        assert_eq!(rows[0].0.display_name, "editor.exe");
-        assert_eq!(rows[0].0.processes[0].id, 30);
-        assert_eq!(rows[0].0.processes[1].id, 20);
-        assert_eq!(rows[1].0.display_name, "worker.exe");
-    }
-
-    #[test]
-    fn process_list_sort_orders_groups_by_policy_column_value() {
-        let processes = vec![
-            ProcessInfo {
-                id: 1,
-                parent_id: None,
-                name: "editor.exe".to_owned(),
-            },
-            ProcessInfo {
-                id: 2,
-                parent_id: None,
-                name: "worker.exe".to_owned(),
-            },
-        ];
-        let groups = process_list_groups(&processes);
-        let mut low = default_process_policy_summary();
-        low.process_priority = "Idle".to_owned();
-        let mut high = default_process_policy_summary();
-        high.process_priority = "Normal".to_owned();
-        let rows = process_list_sorted_rows(
-            groups,
-            vec![high, low],
-            ProcessListSort {
-                column: ProcessListSortColumn::Column(ProcessListColumn::ProcessPriority),
-                direction: ProcessListSortDirection::Ascending,
-            },
-        );
-
-        assert_eq!(rows[0].0.display_name, "worker.exe");
-        assert_eq!(rows[1].0.display_name, "editor.exe");
-    }
-
-    #[test]
-    fn process_list_policy_value_active_tracks_state_and_custom_values() {
-        assert!(process_list_policy_value_active("Include", false));
-        assert!(process_list_policy_value_active("Include (50%)", false));
-        assert!(!process_list_policy_value_active("Exclude", true));
-        assert!(!process_list_policy_value_active(
-            process_list_default_label().as_str(),
-            true
-        ));
-        assert!(!process_list_policy_value_active("Balanced", false));
-        assert!(process_list_policy_value_active("Balanced", true));
-    }
-
-    #[test]
-    fn process_list_split_policy_value_parses_foreground_background_pairs() {
-        assert_eq!(
-            process_list_split_policy_value("Normal / Very low"),
-            Some(("Normal", "Very low"))
-        );
-        assert_eq!(
-            process_list_split_policy_value("  Above normal / Idle  "),
-            Some(("Above normal", "Idle"))
-        );
-        assert_eq!(process_list_split_policy_value("Default"), None);
-    }
-
-    #[test]
-    fn process_list_policy_cell_editing_respects_row_editability() {
-        assert!(!process_list_policy_cell_editable(
-            true,
-            ProcessListColumn::ProcessPriority
-        ));
-        assert!(!process_list_policy_cell_editable(
-            false,
-            ProcessListColumn::ProcessPriority
-        ));
-        assert!(!process_list_policy_cell_editable(
-            true,
-            ProcessListColumn::CoreSteering
-        ));
-    }
-
-    #[test]
-    fn process_policy_summary_matches_exact_process_rule() {
-        let mut settings = Settings::default();
-        settings.core_steering.enabled = true;
-        settings.core_steering.rules.push(CoreSteeringRule {
-            enabled: true,
-            mode: CoreSteeringMode::Soft,
-            process_name: "Editor.EXE".to_owned(),
-            core_mask: 0b1011,
-        });
-
-        let matching = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(matching.core_steering, "0-1, 3");
-        assert!(matching.uses_custom_rule(ProcessListColumn::CoreSteering));
-
-        let non_matching = process_policy_summary(&settings, &[], "browser.exe");
-        assert_eq!(
-            non_matching.power_plan_foreground,
-            process_list_default_label()
-        );
-        assert_eq!(
-            non_matching.power_plan_running,
-            process_list_default_label()
-        );
-        assert_eq!(non_matching.core_steering, default_core_steering_label());
-        assert_eq!(
-            non_matching.process_priority,
-            process_priority_setting_label(ProcessPrioritySetting::Default)
-        );
-        assert!(!non_matching.uses_custom_rule(ProcessListColumn::CoreSteering));
-    }
-
-    #[test]
-    fn process_policy_summary_reports_priority_policy_values() {
-        let mut settings = Settings::default();
-        settings.io_priority.enabled = true;
-        settings.gpu_priority.enabled = true;
-        settings.memory_priority.enabled = true;
-
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-
-        assert_eq!(
-            summary.io_priority,
-            io_priority_policy_label(&settings.io_priority)
-        );
-        assert_eq!(
-            summary.gpu_priority,
-            gpu_priority_policy_label(&settings.gpu_priority)
-        );
-        assert_eq!(
-            summary.memory_priority,
-            memory_priority_policy_label(&settings.memory_priority)
-        );
-    }
-
-    #[test]
-    fn process_policy_summary_reports_process_rule_columns() {
-        let mut settings = Settings::default();
-        settings.by_foreground.enabled = true;
-        settings.by_foreground.rules.push(ByForegroundRule {
-            enabled: true,
-            name: "Editor".to_owned(),
-            process_name: "editor.exe".to_owned(),
-            power_plan_guid: Some("balanced-guid".to_owned()),
-        });
-        settings.by_running_app.enabled = true;
-        settings.by_running_app.rules.push(ByRunningAppRule {
-            enabled: true,
-            name: "Editor".to_owned(),
-            process_name: "editor.exe".to_owned(),
-            power_plan_guid: Some("performance-guid".to_owned()),
-        });
-        settings.core_limiter.enabled = true;
-        settings.core_limiter.rules.push(CoreLimiterRule {
-            enabled: true,
-            process_name: "editor.exe".to_owned(),
-            threshold_percent: 80,
-            sustain_seconds: 5,
-            cooldown_seconds: 30,
-            max_logical_processors: 50,
-        });
-        settings.app_suspension.enabled = true;
-        settings
-            .app_suspension
-            .suspendable_apps
-            .push(AppSuspensionRule {
-                enabled: true,
-                process_name: "editor.exe".to_owned(),
-                network_wake_enabled: true,
-                audio_wake_enabled: true,
-                network_download_threshold_bytes: 1,
-                network_download_threshold_unit: NetworkThresholdUnit::Bytes,
-                network_upload_threshold_bytes: 0,
-                network_upload_threshold_unit: NetworkThresholdUnit::Bytes,
-            });
-        settings.timer_resolution.enabled = true;
-        settings.timer_resolution.rules.push(TimerResolutionRule {
-            enabled: true,
-            process_name: "editor.exe".to_owned(),
-            desired_100ns: 10_000,
-        });
-        let plans = vec![
-            PowerPlan {
-                guid: "balanced-guid".to_owned(),
-                name: "Balanced".to_owned(),
-                active: false,
-            },
-            PowerPlan {
-                guid: "performance-guid".to_owned(),
-                name: "Performance".to_owned(),
-                active: false,
-            },
-        ];
-
-        let summary = process_policy_summary(&settings, &plans, "editor.exe");
-
-        assert_eq!(summary.power_plan_foreground, "Balanced");
-        assert_eq!(summary.power_plan_running, "Performance");
-        assert_eq!(
-            summary.core_limiter,
-            process_list_include_value_label("50%")
-        );
-        assert_eq!(summary.app_suspension, process_list_include_label());
-        assert_eq!(summary.timer_resolution, "1.00 ms");
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanForeground));
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanRunning));
-        assert!(summary.uses_custom_rule(ProcessListColumn::CoreLimiter));
-        assert!(summary.uses_custom_rule(ProcessListColumn::AppSuspension));
-        assert!(summary.uses_custom_rule(ProcessListColumn::TimerResolution));
-    }
-
-    #[test]
-    fn process_policy_summary_reports_include_exclude_columns() {
-        let mut settings = Settings::default();
-        settings
-            .background_efficiency
-            .custom_rules
-            .push(new_background_efficiency_rule("editor.exe"));
-        settings
-            .memory_trim
-            .exclusions
-            .push(new_process_exclusion_rule("editor.exe"));
-        settings
-            .app_suspension
-            .suspendable_apps
-            .push(AppSuspensionRule {
-                enabled: true,
-                process_name: "editor.exe".to_owned(),
-                network_wake_enabled: true,
-                audio_wake_enabled: true,
-                network_download_threshold_bytes: 1,
-                network_download_threshold_unit: NetworkThresholdUnit::Bytes,
-                network_upload_threshold_bytes: 0,
-                network_upload_threshold_unit: NetworkThresholdUnit::Bytes,
-            });
-
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-
-        assert_eq!(summary.background_efficiency, process_list_exclude_label());
-        assert_eq!(summary.core_limiter, process_list_exclude_label());
-        assert_eq!(summary.memory_trim, process_list_exclude_label());
-        assert_eq!(summary.app_suspension, process_list_include_label());
-        assert_eq!(summary.timer_resolution, process_list_default_label());
-        assert!(summary.uses_custom_rule(ProcessListColumn::BackgroundEfficiency));
-        assert!(!summary.uses_custom_rule(ProcessListColumn::CoreLimiter));
-        assert!(summary.uses_custom_rule(ProcessListColumn::MemoryTrim));
-        assert!(summary.uses_custom_rule(ProcessListColumn::AppSuspension));
-        assert!(!summary.uses_custom_rule(ProcessListColumn::TimerResolution));
-    }
-
-    #[test]
-    fn process_policy_summary_reports_priority_exclusions_as_exclude() {
-        let mut settings = Settings::default();
-        settings
-            .io_priority
-            .exclusions
-            .push(new_process_exclusion_rule("editor.exe"));
-        settings
-            .gpu_priority
-            .exclusions
-            .push(new_process_exclusion_rule("editor.exe"));
-        settings
-            .memory_priority
-            .exclusions
-            .push(new_process_exclusion_rule("editor.exe"));
-
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-
-        assert_eq!(summary.io_priority, process_list_exclude_label());
-        assert_eq!(summary.gpu_priority, process_list_exclude_label());
-        assert_eq!(summary.memory_priority, process_list_exclude_label());
-        assert!(summary.uses_custom_rule(ProcessListColumn::IoPriority));
-        assert!(summary.uses_custom_rule(ProcessListColumn::GpuPriority));
-        assert!(summary.uses_custom_rule(ProcessListColumn::MemoryPriority));
-    }
-
-    #[test]
-    fn process_list_rule_edit_helpers_update_process_overrides() {
-        let mut settings = Settings::default();
-
-        set_foreground_power_plan_override(
-            &mut settings.by_foreground,
-            "Editor.EXE",
-            Some("balanced-guid".to_owned()),
-        );
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(summary.power_plan_foreground, "balanced-guid");
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanForeground));
-
-        set_foreground_power_plan_override(&mut settings.by_foreground, "editor.exe", None);
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(summary.power_plan_foreground, process_list_default_label());
-        assert!(!summary.uses_custom_rule(ProcessListColumn::PowerPlanForeground));
-
-        set_core_limiter_override(&mut settings.core_limiter, "editor.exe", Some(50));
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(
-            summary.core_limiter,
-            process_list_include_value_label("50%")
-        );
-        assert!(summary.uses_custom_rule(ProcessListColumn::CoreLimiter));
-
-        set_core_limiter_override(&mut settings.core_limiter, "editor.exe", None);
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(summary.core_limiter, process_list_exclude_label());
-        assert!(!summary.uses_custom_rule(ProcessListColumn::CoreLimiter));
-    }
-
-    #[test]
-    fn process_list_rule_edit_helpers_update_timer_overrides() {
-        let mut settings = Settings::default();
-
-        set_timer_resolution_override(&mut settings.timer_resolution, "editor.exe", Some(20_000));
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(
-            summary.timer_resolution,
-            timer_resolution::format_resolution_ms(20_000)
-        );
-        assert!(summary.uses_custom_rule(ProcessListColumn::TimerResolution));
-
-        set_timer_resolution_override(&mut settings.timer_resolution, "editor.exe", None);
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-        assert_eq!(summary.timer_resolution, process_list_default_label());
-        assert!(!summary.uses_custom_rule(ProcessListColumn::TimerResolution));
-    }
-
-    #[test]
-    fn process_policy_summary_reports_default_power_plan_when_unset() {
-        let mut settings = Settings::default();
-        settings.by_foreground.enabled = true;
-        settings.by_foreground.rules.push(ByForegroundRule {
-            enabled: true,
-            name: "Editor".to_owned(),
-            process_name: "editor.exe".to_owned(),
-            power_plan_guid: None,
-        });
-        settings.by_running_app.enabled = true;
-        settings.by_running_app.rules.push(ByRunningAppRule {
-            enabled: true,
-            name: "Editor".to_owned(),
-            process_name: "editor.exe".to_owned(),
-            power_plan_guid: None,
-        });
-
-        let summary = process_policy_summary(&settings, &[], "editor.exe");
-
-        assert_eq!(summary.power_plan_foreground, process_list_default_label());
-        assert_eq!(summary.power_plan_running, process_list_default_label());
-        assert_eq!(summary.timer_resolution, process_list_default_label());
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanForeground));
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanRunning));
-        assert!(!summary.uses_custom_rule(ProcessListColumn::TimerResolution));
-    }
-
-    #[test]
-    fn process_policy_summary_reports_configured_rules_when_feature_disabled() {
-        let mut settings = Settings::default();
-        settings.by_foreground.enabled = false;
-        settings.by_foreground.rules.push(ByForegroundRule {
-            enabled: true,
-            name: "Editor".to_owned(),
-            process_name: "editor.exe".to_owned(),
-            power_plan_guid: Some("balanced-guid".to_owned()),
-        });
-        settings.by_running_app.enabled = false;
-        settings.by_running_app.rules.push(ByRunningAppRule {
-            enabled: true,
-            name: "Editor".to_owned(),
-            process_name: "editor.exe".to_owned(),
-            power_plan_guid: Some("performance-guid".to_owned()),
-        });
-        settings.core_limiter.enabled = false;
-        settings.core_limiter.rules.push(CoreLimiterRule {
-            enabled: true,
-            process_name: "editor.exe".to_owned(),
-            threshold_percent: 80,
-            sustain_seconds: 5,
-            cooldown_seconds: 30,
-            max_logical_processors: 25,
-        });
-        settings.timer_resolution.enabled = false;
-        settings.timer_resolution.rules.push(TimerResolutionRule {
-            enabled: true,
-            process_name: "editor.exe".to_owned(),
-            desired_100ns: 10_000,
-        });
-        let plans = vec![
-            PowerPlan {
-                guid: "balanced-guid".to_owned(),
-                name: "Balanced".to_owned(),
-                active: false,
-            },
-            PowerPlan {
-                guid: "performance-guid".to_owned(),
-                name: "Performance".to_owned(),
-                active: false,
-            },
-        ];
-
-        let summary = process_policy_summary(&settings, &plans, "editor.exe");
-
-        assert_eq!(summary.power_plan_foreground, "Balanced");
-        assert_eq!(summary.power_plan_running, "Performance");
-        assert_eq!(
-            summary.core_limiter,
-            process_list_include_value_label("25%")
-        );
-        assert_eq!(summary.timer_resolution, "1.00 ms");
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanForeground));
-        assert!(summary.uses_custom_rule(ProcessListColumn::PowerPlanRunning));
-        assert!(summary.uses_custom_rule(ProcessListColumn::CoreLimiter));
-        assert!(summary.uses_custom_rule(ProcessListColumn::TimerResolution));
-    }
-
-    #[test]
-    fn cpu_mask_formatter_uses_ranges() {
-        assert_eq!(format_cpu_mask(0), t!("common.none").to_string());
-        assert_eq!(format_cpu_mask(0b1111), "0-3");
-        assert_eq!(format_cpu_mask(0b101101), "0, 2-3, 5");
-    }
-
-    #[test]
-    fn no_smt_mask_selects_one_logical_cpu_per_physical_core() {
-        let processors = vec![
-            LogicalProcessorInfo {
-                index: 0,
-                core_index: 0,
-                kind: LogicalProcessorKind::Standard,
-                efficiency_class: 0,
-            },
-            LogicalProcessorInfo {
-                index: 1,
-                core_index: 0,
-                kind: LogicalProcessorKind::Standard,
-                efficiency_class: 0,
-            },
-            LogicalProcessorInfo {
-                index: 2,
-                core_index: 1,
-                kind: LogicalProcessorKind::Standard,
-                efficiency_class: 0,
-            },
-            LogicalProcessorInfo {
-                index: 3,
-                core_index: 1,
-                kind: LogicalProcessorKind::Standard,
-                efficiency_class: 0,
-            },
-        ];
-
-        assert_eq!(core_steering_processors_no_smt_mask(&processors), 0b0101);
-    }
-
-    #[test]
-    fn topology_aware_core_toggle_keeps_one_available_cpu_selected() {
-        let mut mask = (1_u64 << 63) | 0b0001;
-        toggle_affinity_core_with_available_mask(&mut mask, 0, 0b0011);
-
-        assert_eq!(mask, 0b0001);
-
-        toggle_affinity_core_with_available_mask(&mut mask, 1, 0b0011);
-        assert_eq!(mask, 0b0011);
-
-        toggle_affinity_core_with_available_mask(&mut mask, 0, 0b0011);
-        assert_eq!(mask, 0b0010);
-    }
-
-    #[test]
-    fn new_core_steering_rules_default_to_soft_cpu_sets() {
-        let rule = new_core_steering_rule("game.exe");
-
-        assert_eq!(rule.mode, CoreSteeringMode::Soft);
-    }
-
-    #[test]
-    fn csv_escape_quotes_fields_with_special_characters() {
-        let mut escaped = String::new();
-        push_csv_field(&mut escaped, "two,parts");
-        assert_eq!(escaped.len(), 11);
-        assert_eq!(csv_escape("plain"), "plain");
-        assert_eq!(csv_escape("two,parts"), "\"two,parts\"");
-        assert_eq!(csv_escape("quoted \"value\""), "\"quoted \"\"value\"\"\"");
-        assert_eq!(csv_escape("line\r\nbreak"), "\"line\r\nbreak\"");
-    }
-
-    #[test]
-    fn timer_resolution_input_accepts_decimal_milliseconds() {
-        assert_eq!(
-            parse_timer_resolution_input_100ns("1", 10_000, 160_000),
-            Some(10_000)
-        );
-        assert_eq!(
-            parse_timer_resolution_input_100ns("0.5 ms", 10_000, 160_000),
-            Some(10_000)
-        );
-        assert_eq!(
-            parse_timer_resolution_input_100ns("15.625 MS", 10_000, 160_000),
-            Some(160_000)
-        );
-    }
-
-    #[test]
-    fn timer_resolution_input_clamps_to_supported_range() {
-        assert_eq!(
-            parse_timer_resolution_input_100ns("0.1", 10_000, 160_000),
-            Some(10_000)
-        );
-        assert_eq!(
-            parse_timer_resolution_input_100ns("1000", 10_000, 160_000),
-            Some(160_000)
-        );
-    }
-
-    #[test]
-    fn action_log_entries_export_as_csv() {
-        let entries = vec![ActionLogEntry {
-            sequence: 7,
-            timestamp_epoch_ms: 1_700_000_000_000,
-            feature: ActionLogFeature::CoreLimiter,
-            process_id: Some(42),
-            process_name: "worker.exe".to_owned(),
-            action: ActionLogAction::Fail,
-            result: ActionLogResult::Failed,
-            reason: "Restart failed, access denied".to_owned(),
-        }];
-
-        let csv = action_log_entries_to_csv(&entries);
-
-        assert!(csv.starts_with(
-            "sequence,timestamp,feature,process_id,process_name,action,result,reason\r\n"
-        ));
-        assert!(csv.contains(
-            ",Core Limiter,42,worker.exe,Fail,Failed,\"Restart failed, access denied\"\r\n"
-        ));
-    }
-
-    #[test]
-    fn action_log_filtering_matches_result_and_feature() {
-        let entries = vec![
-            ActionLogEntry {
-                sequence: 1,
-                timestamp_epoch_ms: 1_700_000_000_000,
-                feature: ActionLogFeature::CoreLimiter,
-                process_id: Some(42),
-                process_name: "worker.exe".to_owned(),
-                action: ActionLogAction::Fail,
-                result: ActionLogResult::Failed,
-                reason: "restart failed".to_owned(),
-            },
-            ActionLogEntry {
-                sequence: 2,
-                timestamp_epoch_ms: 1_700_000_000_100,
-                feature: ActionLogFeature::GpuPriority,
-                process_id: Some(43),
-                process_name: "game.exe".to_owned(),
-                action: ActionLogAction::Apply,
-                result: ActionLogResult::Applied,
-                reason: "priority applied".to_owned(),
-            },
-        ];
-
-        let filtered_entries = action_log_filtered_entries(
-            &entries,
-            ActionLogResultFilter::Failed,
-            ActionLogFeatureFilter::Feature(ActionLogFeature::CoreLimiter),
-        );
-
-        assert_eq!(filtered_entries.len(), 1);
-        assert_eq!(filtered_entries[0].sequence, 1);
-    }
-
-    #[test]
-    fn action_log_page_count_rounds_up() {
-        assert_eq!(action_log_page_count(0), 0);
-        assert_eq!(action_log_page_count(1), 1);
-        assert_eq!(action_log_page_count(ACTION_LOG_PAGE_SIZE), 1);
-        assert_eq!(action_log_page_count(ACTION_LOG_PAGE_SIZE + 1), 2);
-    }
-
-    #[test]
-    fn processor_power_slider_pairs_ac_and_battery_controls() {
-        assert_eq!(
-            ProcessorPowerSlider::AcCoreParkingMin.paired_power_source(),
-            ProcessorPowerSlider::DcCoreParkingMin
-        );
-        assert_eq!(
-            ProcessorPowerSlider::AcPerformanceMin.paired_power_source(),
-            ProcessorPowerSlider::DcPerformanceMin
-        );
-        assert_eq!(
-            ProcessorPowerSlider::AcPerformanceMax.paired_power_source(),
-            ProcessorPowerSlider::DcPerformanceMax
-        );
-        assert_eq!(
-            ProcessorPowerSlider::AcBoostPolicy.paired_power_source(),
-            ProcessorPowerSlider::DcBoostPolicy
-        );
-        assert_eq!(
-            ProcessorPowerSlider::DcCoreParkingMin.paired_power_source(),
-            ProcessorPowerSlider::AcCoreParkingMin
-        );
-        assert_eq!(
-            ProcessorPowerSlider::DcBoostPolicy.paired_power_source(),
-            ProcessorPowerSlider::AcBoostPolicy
-        );
     }
 
     #[test]
